refactor: split node_catalog.rs into a submodule directory
The 830-line node_catalog.rs was the largest hand-written module. Split it
along its existing concept boundaries into node_catalog/, one file per job:

  mod.rs      - types (NodeKind, CatalogEntry, NodeCatalog) + the build() driver
  classify.rs - per-world / per-entry NodeKind classification
  pipes.rs    - pipe-pair matching and A-side detection
  naming.rs   - human-readable entry names ("1-1", "3F2", "7-P1", ...)
  tests.rs    - catalog build invariants + the --ignored dump

Pure code move: function bodies are byte-identical extractions, only import
paths and cross-module visibility (pub(super)) were adjusted. Verified the
generated IPS is byte-for-byte identical to pre-refactor across 5 seeds, a
kitchen-sink flag combo, and the --include-beta-stages path.

(palette_variants.rs is larger but machine-generated by gen_palette_variants.py,
so splitting it would break regeneration.)
